添加了获取摄像头url功能，更新了一键删除

+++ database.py
@@ -1366,42 +1366,49 @@ def get_report(user_id: int):
         db.close()
 
 
-
 def delete_all_cam(uid):
     try:
         db = pymysql.connect(host="zrp.cool", user="CAMotion", passwd="M4RpMGAKFhBBARGx", db="CAMotion",
                              port=3306,
                              charset='utf8')
+
         cursor = db.cursor()
-        sql = 'select * from cams where uid="%d";' % uid
+        sql = 'select * from users where id="%d";' % uid
         cursor.execute(sql)
-        results = cursor.fetchall()
-
-        if results:
-            for row in results:
-                cid = row[0]
-                cursor = db.cursor()
-                sql = 'select * from logs where cid=%d;' % cid
-                cursor.execute(sql)
-                results_per_camera = cursor.fetchall()
-                if results_per_camera:
-                    sql = 'delete from logs where cid=%d;' % cid
-                    cursor.execute(sql)
-                    db.commit()
-
-            sql = 'delete from cams where uid=%d;' % uid
+        results_user = cursor.fetchall()
+        if results_user:
+            cursor = db.cursor()
+            sql = 'select * from cams where uid="%d";' % uid
             cursor.execute(sql)
-            db.commit()
-            cursor.close()
-            result = {
-                "status": "Success",
-            }
-            return result
+            results = cursor.fetchall()
+            if results:
+                for row in results:
+                    cid = row[0]
+                    cursor = db.cursor()
+                    sql = 'select * from logs where cid=%d;' % cid
+                    cursor.execute(sql)
+                    results_per_camera = cursor.fetchall()
+                    if results_per_camera:
+                        sql = 'delete from logs where cid=%d;' % cid
+                        cursor.execute(sql)
+                        db.commit()
 
+                sql = 'delete from cams where uid=%d;' % uid
+                cursor.execute(sql)
+                db.commit()
+                cursor.close()
+                result = {
+                    "status": "Success",
+                }
+            else :
+                result = {
+                    "status": "Failed",
+                    "message": "Cannot found the user's camera"
+                }
         else:
             result = {
                 "status": "Failed",
-                "message": "Cannot found the user's camera"
+                "message": "Cannot found the user"
             }
         return result
 
@@ -1413,13 +1420,13 @@ def delete_all_cam(uid):
             "message": "Data too long"
         }
         return result
-    except pymysql.err.OperationalError:
-        traceback.print_exc()
-        result = {
-            "status": "Failed",
-            "message": "OperationalError"
-        }
-        return result
+    # except pymysql.err.OperationalError:
+    #     traceback.print_exc()
+    #     result = {
+    #         "status": "Failed",
+    #         "message": "OperationalError"
+    #     }
+    #     return result
     except IndexError:
         traceback.print_exc()
         result = {
@@ -1441,6 +1448,50 @@ def delete_all_cam(uid):
             "message": "local variable referenced before assignment"
         }
         return result
+    except:
+        traceback.print_exc()
+        f = open("exceptionLog.txt", 'a')
+        traceback.print_exc(file=f)
+        f.flush()
+        f.close()
+        db.rollback()
+    finally:
+        db.close()
+
+
+def get_url(uid, cid):
+    try:
+        db = pymysql.connect(host="zrp.cool", user="CAMotion", passwd="M4RpMGAKFhBBARGx", db="CAMotion",
+                             port=3306,
+                             charset='utf8')
+        cursor = db.cursor()
+        sql = 'select * from cams where uid="%d";' % uid
+        cursor.execute(sql)
+        results = cursor.fetchall()
+
+        if results:
+            cursor = db.cursor()
+            sql = 'select * from cams where id="%d";' % cid
+            cursor.execute(sql)
+            results_per_camera = cursor.fetchall()
+            if results_per_camera:
+                for row in results_per_camera:
+                    url = row[2]
+                    result = {
+                        "url": url
+                    }
+            else:
+                result = {
+                    "message": "Cannot found the camera"
+                }
+        else:
+            result = {
+                "message": "Cannot found the user"
+            }
+        cursor.close()
+        return result
+
+
     except:
         traceback.print_exc()
         f = open("exceptionLog.txt", 'a')
